Update: Radical modularization of Cycle into components which can be improved individually
