+++ src/Cycle.c
@@ -172,104 +172,6 @@ static Decision Cycle_DeriveSubgoals(long currentTime)
     return best_decision;
 }
 
-//Reinforce link between concept a and b (creating it if non-existent)
-static void Cycle_ReinforceLink(Event *a, Event *b)
-{
-    if(a->type != EVENT_TYPE_BELIEF || b->type != EVENT_TYPE_BELIEF)
-    {
-        return;
-    }
-    Term a_term_nop = Narsese_GetPreconditionWithoutOp(&a->term);
-    Concept *A = Memory_FindConceptByTerm(&a_term_nop);
-    Concept *B = Memory_FindConceptByTerm(&b->term);
-    if(A != NULL && B != NULL && A != B)
-    {
-        //temporal induction
-        if(!Stamp_checkOverlap(&a->stamp, &b->stamp))
-        {
-            bool success;
-            Implication precondition_implication = Inference_BeliefInduction(a, b, &success);
-            if(success)
-            {
-                precondition_implication.sourceConcept = A;
-                precondition_implication.sourceConceptId = A->id;
-                if(precondition_implication.truth.confidence >= MIN_CONFIDENCE)
-                {
-                    bool success;
-                    Term general_implication_term = IntroduceImplicationVariables(precondition_implication.term, &success);
-                    if(success && Variable_hasVariable(&general_implication_term, true, true, false))
-                    {
-                        NAL_DerivedEvent(general_implication_term, OCCURRENCE_ETERNAL, precondition_implication.truth, precondition_implication.stamp, currentTime, 1, 1, precondition_implication.occurrenceTimeOffset, NULL, 0);
-                    }
-                    int operationID = Narsese_getOperationID(&a->term);
-                    IN_DEBUG( fputs("Formed implication: ", stdout); Narsese_PrintTerm(&precondition_implication.term); Truth_Print(&precondition_implication.truth); puts("\n"); )
-                    Implication *revised_precon = Table_AddAndRevise(&B->precondition_beliefs[operationID], &precondition_implication);
-                    if(revised_precon != NULL)
-                    {
-                        revised_precon->creationTime = currentTime; //for evaluation
-                        revised_precon->sourceConcept = A;
-                        revised_precon->sourceConceptId = A->id;
-                        /*IN_DEBUG( fputs("REVISED pre-condition implication: ", stdout); Implication_Print(revised_precon); )*/
-                        Memory_printAddedImplication(&revised_precon->term, &revised_precon->truth, false, revised_precon->truth.confidence > precondition_implication.truth.confidence);
-                    }
-                }
-            }
-        }
-    }
-}
-
-void Cycle_CorrelateEvents(long currentTime)
-{
-    //1. process newest event
-    if(belief_events.itemsAmount > 0)
-    {
-        //form concepts for the sequences of different length
-        for(int len=0; len<MAX_SEQUENCE_LEN; len++)
-        {
-            Event *toProcess = FIFO_GetNewestSequence(&belief_events, len);
-            if(toProcess != NULL && !toProcess->processed && toProcess->type != EVENT_TYPE_DELETED)
-            {
-                assert(toProcess->type == EVENT_TYPE_BELIEF, "A different event type made it into belief events!");
-                Memory_ProcessNewBeliefEvent(toProcess, currentTime, 1.0, 0, false, true, false, false);
-                Event postcondition = *toProcess;
-                //Mine for <(&/,precondition,operation) =/> postcondition> patterns in the FIFO:
-                if(len == 0) //postcondition always len1
-                {
-                    int op_id = Narsese_getOperationID(&postcondition.term);
-                    Decision_AssumptionOfFailure(op_id, currentTime); //collection of negative evidence, new way
-                    //build link between internal derivations and external event to explain it:
-                    for(int k=0; k<beliefsSelectedCnt; k++)
-                    {
-                        if(selectedBeliefs[k].occurrenceTime < postcondition.occurrenceTime)
-                        {
-                            Cycle_ReinforceLink(&selectedBeliefs[k], &postcondition);
-                        }
-                    }
-                    for(int k=1; k<belief_events.itemsAmount; k++)
-                    {
-                        for(int len2=0; len2<MAX_SEQUENCE_LEN; len2++)
-                        {
-                            Event *precondition = FIFO_GetKthNewestSequence(&belief_events, k, len2);
-                            if(len2 > 0)
-                            {
-                                Event *potential_op = FIFO_GetKthNewestSequence(&belief_events, k+len2, 0);
-                                if(potential_op != NULL && potential_op->type != EVENT_TYPE_DELETED && Narsese_isOperation(&potential_op->term))
-                                {
-                                    break;
-                                }
-                            }
-                            if(precondition != NULL && precondition->type != EVENT_TYPE_DELETED)
-                            {
-                                Cycle_ReinforceLink(precondition, &postcondition);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
 void Cycle_GoalReasoning(long currentTime)
 {
     //process goals, allow reasoning into the future by propagating spikes from goals back to potential current events
